# SSD performance optimization and benchmark script