## Description ##
In this PR:

- An inference benchmark script for SSD is added
- Operator MultiBoxDetection is optimized to improve SSD performance on CPU
- I tried to add a unit test for the consistency check between cpu and gpu but failed, since MultiBoxDetection on gpu cannot generate stable results for different run. I have raised a issue for this #10316. This code has past the consistency check with original cpu implementation.

`python benchmark_score.py`
SSD-VGG16 with 20 classes and shape (3, 300, 300), img/sec

bs | master | +mkldnn | +mkldnn+this PR
:--: | :--: | :--: | :--:
1 | 2.02 | 7.64 | 23.77
2 | 2.10 | 5.28 | 15.54
4 | 2.26 | 6.80 | 23.08
8 | 2.40 | 8.03 | 28.78
16 | 2.80 | 9.11 | 33.14
32 | 2.82 | 8.80 | 32.81

`python benchmark_score.py --network resnet50`
SSD-ResNet50 with 20 classes and shape (3, 300, 300), img/sec

bs | master | +mkldnn | +mkldnn+this PR
:--: | :--: | :--: | :--:
1 | 8.32 | 30.94 | 40.58
2 | 8.58 | 28.71 | 37.65
4 | 8.69 | 39.02 | 54.04
8 | 8.59 | 47.72 | 68.90
16 | 8.54 | 47.64 | 77.39
32 | 8.79 | 54.74 | 89.62



## Checklist ##
### Essentials ###
Please feel free to remove inapplicable items for your PR.
- [ ] The PR title starts with [MXNET-$JIRA_ID], where $JIRA_ID refers to the relevant [JIRA issue](https://issues.apache.org/jira/projects/MXNET/issues) created (except PRs with tiny changes)
- [ ] Changes are complete (i.e. I finished coding on this PR)
- [ ] All changes have test coverage:
- Unit tests are added for small changes to verify correctness (e.g. adding a new operator)
- Nightly tests are added for complicated/long-running ones (e.g. changing distributed kvstore)
- Build tests will be added for build configuration changes (e.g. adding a new build option with NCCL)
- [ ] Code is well-documented: 
- For user-facing API changes, API doc string has been updated. 
- For new C++ functions in header files, their functionalities and arguments are documented. 
- For new examples, README.md is added to explain the what the example does, the source of the dataset, expected performance on test set and reference to the original paper if applicable
- Check the API doc at http://mxnet-ci-doc.s3-accelerate.dualstack.amazonaws.com/PR-$PR_ID/$BUILD_ID/index.html
- [ ] To the my best knowledge, examples are either not affected by this change, or have been fixed to be compatible with this change

### Changes ###
- [ ] Feature1, tests, (and when applicable, API doc)
- [ ] Feature2, tests, (and when applicable, API doc)

## Comments ##
- If this change is a backward incompatible change, why must this change be made.
- Interesting edge cases to note here