@pengzhao-intel 2018-04-10
There's huge improvements for SSD.  
Is it possible to add this one into release 1.2?