@zhreshold 2018-04-10
I hold the same opinion with @kevinthesun , it is not precise to benchmark randomly intialized weights.