@piiswrong 2018-04-10
So this would only help at the beginning of training?