@kevinthesun 2018-04-10
MultiBoxDetection is only used for inference. @zhreshold Is this accurate?