@zhreshold 2018-04-10
I think this should still improve the inference speed, but we would like to see how this affecting the pre-trained models. Apparently not by such a huge margin.