@pengzhao-intel 2018-04-11
@zhreshold @kevinthesun @piiswrong where we can get the pre-trained SSD data?