@juliusshufan 2018-04-11
@kevinthesun Yes, pre-trained model can be used for inference benchmarking, while it still involving dataloader and dataset, the benchmark script appended with this PR is dummy data-based and might be more flexible, I guess it is a potential utility as well. What do y ou think? :)