@kevinthesun 2018-04-14
@juliusshufan Synthetic data is fine as along as pre-trained model is used.