@zhreshold 2018-05-24
LGTM for this PR, just wish the benchmarks could be updated for the pretrained models so we can have better ideas of the perf improvement.
I will merge this now.